deleteFolderコマンドにソフト削除とハード削除を実装

+++ src-tauri/src/types.rs
@@ -37,13 +37,13 @@ pub struct RawDetailMemo {
     pub tags: Option<String>,
 }
 
-#[derive(Debug, Serialize, Deserialize, FromRow)]
+#[derive(Debug, Serialize, Deserialize, FromRow, Clone)]
 #[typeshare]
 pub struct CreateMemoIn {
     /// メモタイトル
     pub title: String,
     /// メモフォルダID
-    pub folder_id: Option<i64>,
+    pub folder_id: Option<i32>,
     /// メモ内容
     pub content: String,
     /// メモタグ
